Update Nine_lives.py
Any creative feedback or suggestions would be greatly appreciated. This is literally the first thing I made and its out of a book...
so if you dont mind giving a brief explanation why I'm stupid as opposed to the degree of which I may be, I'd appreciate that as well. Thanks